+++ src/util/property.ts
@@ -10,7 +10,10 @@ export type PropertiesMap = {
     "statusBarPositionPriority": number | undefined;
     "pluralWebPathSubstrings": string[] | undefined;
     "statusBarMessageClickAction": "Show info message" | "Open tool URL";
-    "inlineBlameMargin": number | undefined;
+    "inlineMessageFormat": string;
+    "inlineMessageNoCommit": string;
+    "inlineMessageEnabled": boolean;
+    "inlineBlameMargin": number;
 }
 
 // getConfiguration has an unfortunate typing that does not